+++ include/pomerol/GFContainer.hpp
@@ -44,8 +44,8 @@ namespace Pomerol {
 class GFContainer : public IndexContainer2<GreensFunction, GFContainer>, public Thermal {
 
 public:
-    /// Matrix elements with magnitudes equal to or below this value are treated as negligible.
-    RealType MatrixElementTolerance = 1e-8;
+    /// Minimal magnitude of the coefficient of a term for it to be taken into account.
+    RealType CoefficientTolerance = 1e-8;
 
     /// Constructor.
     /// \tparam IndexTypes Types of indices carried by the creation and annihilation operators.