+++ include/pomerol/GreensFunction.hpp
@@ -67,8 +67,8 @@ class GreensFunction : public Thermal, public ComputableObject {
     std::vector<GreensFunctionPart> parts;
 
 public:
-    /// Matrix elements with magnitudes equal to or below this value are treated as negligible.
-    RealType MatrixElementTolerance = 1e-8;
+    /// Minimal magnitude of the coefficient of a term for it to be taken into account.
+    RealType CoefficientTolerance = 1e-8;
 
     /// Constructor.
     /// \param[in] S Information about invariant subspaces of the Hamiltonian.