+++ include/pomerol/SusceptibilityPart.hpp
@@ -194,8 +194,8 @@ public:
 private:
     /// A difference in energies with magnitude below this value is treated as zero.
     RealType ReduceResonanceTolerance = 1e-8;
-    /// Matrix elements with magnitudes below this value are treated as negligible.
-    RealType MatrixElementTolerance = 1e-8;
+    /// Minimal magnitude of the coefficient of a term for it to be taken into account.
+    RealType CoefficientTolerance = 1e-8;
 
     // Implementation detail of compute().
     template <bool AComplex, bool BComplex> void computeImpl();